feat(#226): Split BytecodeInstruction on CommandInstruction and MarkLabelInstruction
